+++ glutin/src/platform/windows/mod.rs
@@ -2,7 +2,7 @@
 
 use crate::{
     Api, ContextCurrentState, ContextError, CreationError, GlAttributes,
-    GlRequest, NotCurrentContext, PixelFormat, PixelFormatRequirements,
+    GlRequest, NotCurrent, PixelFormat, PixelFormatRequirements,
 };
 
 use crate::api::egl::{Context as EglContext, NativeDisplay, EGL, SurfaceType as EglSurfaceType};
@@ -305,7 +305,7 @@ pub trait RawContextExt {
     unsafe fn build_raw_context(
         self,
         hwnd: *mut raw::c_void,
-    ) -> Result<crate::RawContext<NotCurrentContext>, CreationError>
+    ) -> Result<crate::RawContext<NotCurrent>, CreationError>
     where
         Self: Sized;
 }
@@ -317,7 +317,7 @@ impl<'a, T: ContextCurrentState> RawContextExt
     unsafe fn build_raw_context(
         self,
         hwnd: *mut raw::c_void,
-    ) -> Result<crate::RawContext<NotCurrentContext>, CreationError>
+    ) -> Result<crate::RawContext<NotCurrent>, CreationError>
     where
         Self: Sized,
     {