+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Unreleased
 
+ - **Breaking**: Removed `DisplayLost` variant to `ContextError`.
+ - **Breaking**: Renamed `NotCurrentContext` to `NotCurrent`.
+ - **Breaking**: Renamed `PossiblyCurrentContext` to `PossiblyCurrent`.
+ - Added `treat_as_current` function.
+
 # Version 0.21.0-rc1 (2019-03-09)
 
  - **Breaking:** Replaced `CreationErrorPair` enum variant with `CreationErrors`.